requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.7.0
cycler==0.10.0
Cython==0.29.4
dask==1.1.1
decorator==4.3.2
dominate==2.3.5
ffnet==0.8.4
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
hdf5storage==0.1.15
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
opencv-python==4.0.0.21
Pillow==5.4.1
pkg-resources==0.0.0
protobuf==3.6.1
pydensecrf==1.0rc2
pyparsing==2.3.1
python-dateutil==2.7.5
PyWavelets==1.0.1
pyzmq==17.1.2
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
spacepy==0.1.6
tensorboard==1.12.2
tensorboardX==1.6
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.30.0
urllib3==1.24.1
visdom==0.1.8.8
websocket-client==0.54.0
Werkzeug==0.14.1
